requirements_new.txt
tensorflow>=2.9.0
keras>=2.9.0
numpy>=1.22.0
opencv-python>=4.6.0.66
scikit-image>=0.19.0
scikit-learn>=1.1.0
pandas>=1.4.0
matplotlib>=3.5.0
seaborn>=0.12.0
streamlit>=1.15.0
Pillow>=9.0.0
tqdm>=4.64.0
scipy>=1.8.0
pywavelets>=1.3.0
joblib>=1.1.0
